[step3] 의존성 리팩토링 리뷰요청 드립니다. (#216)
* refactor : 2단계 피드백 및 리팩토링

* refactor: @OneToMany 관계 cascade 변경

* refactor: Order <-> Menu 의존성 제거

* fix: 불필요한 테스트 제거

* refactor: 테이블 주문 검증 로직 리팩토링

+++ src/main/java/kitchenpos/tablegroup/application/TableGroupService.java
@@ -1,5 +1,6 @@
 package kitchenpos.tablegroup.application;
 
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableRepository;
 import kitchenpos.ordertable.dto.OrderTableRequest;
@@ -10,6 +11,8 @@ import kitchenpos.tablegroup.dto.TableGroupResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 
 @Service
@@ -33,12 +36,13 @@ public class TableGroupService {
     public void ungroup(final Long tableGroupId) {
         TableGroup tableGroup = tableGroupRepository.findById(tableGroupId)
                 .orElseThrow(() -> new IllegalArgumentException("단체 지정된 그룹이 없습니다."));
+
+        if (orderTableRepository.findAllByIdIn(
+                tableGroup.getOrderTables(), Arrays.asList(OrderStatus.COOKING.name(), OrderStatus.MEAL.name()))) {
+            throw new IllegalArgumentException();
+        }
+
         tableGroup.ungroupTables();
         tableGroupRepository.delete(tableGroup);
-    }
-
-    private OrderTable findByIdOrderTable(OrderTableRequest orderTableRequest) {
-        return orderTableRepository.findById(orderTableRequest.getId())
-                .orElseThrow(() -> new IllegalArgumentException("주문 테이블을 찾을 수 없습니다."));
     }
 }
